+++ src/lang/Messages.properties
@@ -466,18 +466,22 @@ authentication.method.manual.dialog.error.nosession.text 	= No HTTP Session has 
 authentication.method.manual.dialog.error.title 			= Configuration Error
 authentication.method.manual.field.description				= <html><p><small>* If the HTTP sessions that you are interested in are not present in the selection box, <br/>make sure you have included the <i>Site</i> to which the HTTP sessions correspond<br/> in the proper <b>Context</b>.</small></p><html>
 authentication.method.fb.name								= Form-based Authentication
-authentication.method.fb.field.label.postData				= Login Request POST Data (if any):
-authentication.method.fb.field.label.loginUrl				= Login Form Target URL *:
-authentication.method.fb.field.label.usernameParam			= Username Parameter *:
-authentication.method.fb.field.label.passwordParam			= Password Parameter *:
-authentication.method.fb.field.label.description			= <html><small>The <i>username</i> and <i>password</i> fields will be replaced, during authentication, with the username and password corresponding to application's users.</small><html>
+authentication.method.fb.popup.login.request = {0} : Form-based Auth Login Request
+authentication.method.jb.name = JSON-based Authentication
+authentication.method.jb.popup.login.request = {0} : JSON-based Auth Login Request
+authentication.method.pb.field.label.postData				= Login Request POST Data (if any):
+authentication.method.pb.field.label.postDataRequired = Login Request POST Data *:
+authentication.method.pb.field.label.loginUrl				= Login Form Target URL *:
+authentication.method.pb.field.label.usernameParam			= Username Parameter *:
+authentication.method.pb.field.label.passwordParam			= Password Parameter *:
+authentication.method.pb.field.label.description			= <html><small>The <i>username</i> and <i>password</i> fields will be replaced, during authentication, with the username and password corresponding to application's users.</small><html>
+authentication.method.pb.dialog.error.url.text = The login url you have inserted is not a valid URL.
+authentication.method.pb.dialog.error.postData.text = The POST Data is required.
 authentication.method.fb.credentials.field.label.user		= Username:
 authentication.method.fb.credentials.field.label.pass		= Password:
 authentication.method.fb.dialog.error.nofields.text 		= The username and password form field names have to be configured for this authentication method.
-authentication.method.fb.dialog.error.url.text 				= The login url you have inserted is not a valid URL.
 authentication.method.fb.credentials.dialog.error.user.text	= The username for the current user has to be configured.
 authentication.method.fb.dialog.error.title 				= Configuration Error
-authentication.method.fb.popup.login.request 				= {0} : Form-based Auth Login Request
 authentication.method.http.name								= HTTP/NTLM Authentication
 authentication.method.http.field.label.hostname				= Hostname:
 authentication.method.http.field.label.port					= Port:
